Issue #110 Rename tasks.tsx -> task_list_item.tsx to add clarity
