+++ src/components/tasks/connected_saved_tasks.ts
@@ -1,7 +1,7 @@
 import { Dispatch } from 'redux';
 import { connect } from 'react-redux';
 import { Component, Props, renderSavedTask } from './task_list';
-import { Actions } from './task';
+import { Actions } from './task_list_item';
 import { Store } from '../../application/store';
 import { selectAllSavedTasks } from '../../selectors/tasks';
 import * as stores from '../../stores/tasks';
